automation-agency-office/products/missed_call_lead_capture/requirements.txt
flask>=3.0.0
twilio>=9.0.0
google-api-python-client>=2.120.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.1
gunicorn>=22.0.0
